pyproject.toml
[project]
name = "shift-code-auto-submitter"
version = "0.1.0"
description = "FastAPI backend for discovering and tracking Borderlands SHiFT codes."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.12.3",
    "fastapi>=0.115.0",
    "httpx>=0.27.2",
    "pydantic-settings>=2.6.1",
    "uvicorn[standard]>=0.32.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.3",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
